Moving SpatialSearchResult to a separate file
